Remove reliance on <utility>, which should remove
reliance on libstdc++ specific ABI stuff. We still link to cxxabi stuff,
but clang and gcc both implement IABI much the same, so that should be
decently okay (I hope.) Once we switch to eastl, iirc, it provides its
own forward, so we can use that if desired.

Also removes the need to define the metadata getter if using
plugincommon, you can just define a global of your plugin metadata
and it will just work.